# Meetings index in a component optimization