#### :tophat: What? Why?

This PR adds fragment caching to meetings index of a component performance by fragment caching the _meetings partial render
 * The cache key used is defined in a similar way as in cells with a cache_hash method. This method takes into account:
   * The list of meetings
   * The filter params of the request
   * The current page
   * The cache version of the meetings query (cache_key_with_version does not work because it adds a hash related to the sql query that varies with each request because it uses the current date and time)
   * Cache key with version of component and current_user if present
   * The locale is included by default so it's not included there
 * The cache expires at 10 minutes
 * The PR also adds a touch:true to Follow association with followable. This allows the cache to be expired when changes in meetings followers occur.

##### Metrics

Data obtained from appsignal data of staging applicaction, with a component with 48 meetings and pagination with 20 items per page. Time in ms.

The main improvement is observed in the times of action_view

*Before optimization*

  | Mean | Sample 1 | Sample 2 | Sample 3 | Sample 4 | Sample 5
-- | -- | -- | -- | -- | -- | --
action_view | 1231,4 | 978 | 943 | 1400 | 1920 | 916
active_record | 182 | 126 | 125 | 138 | 388 | 133
action_controller | 69,2 | 59 | 57 | 77 | 96 | 57
active_support | 113 | 90 | 55 | 49 | 282 | 89

*After optimization*


  | Mean | Sample 1 | Sample 2 | Sample 3 | Sample 4 | Sample 5
-- | -- | -- | -- | -- | -- | --
action_view | 335,6 | 348 | 367 | 272 | 333 | 358
active_record | 115 | 112 | 127 | 106 | 106 | 124
action_controller | 118,6 | 285 | 105 | 85 | 58 | 60
active_support | 33,4 | 4 | 16 | 38 | 108 | 1


#### :pushpin: Related Issues

- Related to https://github.com/decidim/decidim/issues/8471

#### Testing

Verify the page loads as before, and maybe in the logs check the cache hit message.

#### :clipboard: Checklist
:rotating_light: Please review the [guidelines for contributing](https://github.com/decidim/decidim/blob/develop/CONTRIBUTING.adoc) to this repository.

- [ ] :question: **CONSIDER** adding a unit test if your PR resolves an issue.
- [ ] :heavy_check_mark: **DO** check open PR's to avoid duplicates.
- [ ] :heavy_check_mark: **DO** keep pull requests small so they can be easily reviewed.
- [ ] :heavy_check_mark: **DO** build locally before pushing.
- [ ] :heavy_check_mark: **DO** make sure tests pass.
- [ ] :heavy_check_mark: **DO** make sure any new changes are documented in `docs/`.
- [ ] :heavy_check_mark: **DO** add and modify seeds if necessary.
- [ ] :heavy_check_mark: **DO** add CHANGELOG upgrade notes if required.
- [ ] :heavy_check_mark: **DO** add to GraphQL API if there are new public fields.
- [ ] :heavy_check_mark: **DO** add link to MetaDecidim if it's a new feature.
- [ ] :x:**AVOID** breaking the continuous integration build.
- [ ] :x:**AVOID** making significant changes to the overall architecture.

### :camera: Screenshots
*Please add screenshots of the changes you're proposing*
![Description](URL)

:hearts: Thank you!