@ferblape 2022-04-30
@ahukkanen you are totally right, and I think some of this user customizations are removed in the new design, where meetings are more compact. 

That said I'm happy to discard this PR if you feel to until the new design is in place to avoid adding more logic layers that can be simplified later.