@ahukkanen 2022-04-30
@ferblape If you asked me, I would discard this PR. And exactly for that reason, we hopefully wouldn't need this kind of stuff in the future, i.e. there needs to be a better way to optimize the performance.

But I'm not the one pulling the cords here, so let's wait for response from @decidim/product .