@andreslucena 2022-05-04
I agree with what you've said, let's discard this. Caching with filtering doesn't work well. 

> @andreslucena https://github.com/orgs/decidim/teams/product Have you agreed on this approach beforehand?

The task that we have pending is just to improve general performance (see #8495). It's tricky because there are some things that probably doesn't make sense to solve as they could be directly related to the UI and the current design, and they will change with the redesign.